+++ InputPassword/index.jsx
@@ -5,12 +5,18 @@ const InputPassword = ({
   input,
   label,
   meta,
+  focused,
+  onFocus,
+  onBlur,
 }) =>
   <Input
     input={input}
     label={label}
     meta={meta}
     type={'password'}
+    focused={focused}
+    onFocus={onFocus}
+    onBlur={onBlur}
   />;
 
 InputPassword.propTypes = {