+++ InputPassword/story.jsx
@@ -30,7 +30,17 @@ storiesOf('InputPassword')
     <InputPassword
       input={input}
       meta={{
-        submitting: true
+        submitting: true,
       }}
+    />
+  ))
+  .add('focused', () => (
+    <InputPassword input={input} focused />
+  ))
+  .add('onFocus + onBlur', () => (
+    <InputPassword
+      input={input}
+      onFocus={action('on-focus')}
+      onBlur={action('on-blur')}
     />
   ));