+++ InputTime/index.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import styles from './style.css';
+import { calculateStyles } from '../lib/utils';
 
 // generate array of numbers (inclusive)
 const genArray = (start, end) => [...Array(end + 1).keys()].slice(start);
@@ -7,20 +7,46 @@ const leftPadTimeUnit = timeUnit => (timeUnit < 10 ? `0${timeUnit}` : timeUnit);
 
 /* eslint-disable react/prop-types */
 
-const renderAmPm = ({ value, onChange, submitting }) =>
-  <select
-    value={value.hours < 12 ? 'am' : 'pm'}
-    disabled={submitting}
-    onChange={e => onChange({
-      ...value,
-      hours: e.target.value === 'am'
-      ? value.hours - 12
-      : value.hours + 12,
-    })}
-  >
-    <option value="am">am</option>
-    <option value="pm">pm</option>
-  </select>;
+const renderAmPm = ({
+  noStyle,
+  onChange,
+  submitting,
+  value,
+}) => {
+  const style = calculateStyles({
+    default: {
+      display: 'inline-flex',
+      marginRight: '0.25rem',
+    },
+    noStyle: {
+      border: 0,
+      background: 'transparent',
+      margin: 0,
+      padding: 0,
+      '-webkit-appearance': 'none',
+      '-moz-appearance': 'none',
+    },
+  }, {
+    noStyle,
+  });
+
+  return (
+    <select
+      disabled={submitting}
+      onChange={e => onChange({
+        ...value,
+        hours: e.target.value === 'AM'
+        ? value.hours - 12
+        : value.hours + 12,
+      })}
+      style={style}
+      value={value.hours < 12 ? 'AM' : 'PM'}
+    >
+      <option value="AM">AM</option>
+      <option value="PM">PM</option>
+    </select>
+  );
+};
 
 /* eslint-enable react/prop-types */
 
@@ -35,55 +61,76 @@ const displayHour = (hour, select24Hours) => {
 };
 
 const InputTime = ({
-  select24Hours,
   input: {
     value,
     onChange,
   },
   meta: {
     submitting,
-  }
+  },
+  noStyle,
+  select24Hours,
 }) => {
+  const style = calculateStyles({
+    default: {
+      display: 'inline-flex',
+      marginRight: '0.25rem',
+    },
+    noStyle: {
+      border: 0,
+      background: 'transparent',
+      marginTop: 0,
+      marginRight: '0.25rem',
+      marginBottom: 0,
+      marginLeft: 0,
+      padding: 0,
+      '-webkit-appearance': 'none',
+      '-moz-appearance': 'none',
+    },
+  }, {
+    noStyle,
+  });
+
   if (!value) {
     value = { hours: 0, minutes: 0 };
   }
   return (
-    <div>
+    <div style={style}>
       <select
-        className={styles['select-with-margin']}
         disabled={submitting}
-        value={value.hours}
         onChange={e => onChange({ ...value, hours: parseInt(e.target.value, 10) })}
+        style={style}
+        value={value.hours}
       >
         {genArray(
           select24Hours || value.hours < 12 ? 0 : 12,
           select24Hours || value.hours > 11 ? 23 : 11,
         ).map(hour =>
-          <option
+          (<option
             key={hour}
             value={hour}
           >
             {leftPadTimeUnit(displayHour(hour, select24Hours))}
-          </option>)
+          </option>))
         }
       </select>
       <select
-        className={!select24Hours ? styles['select-with-margin'] : undefined}
         disabled={submitting}
-        value={value.minutes}
         onChange={e => onChange({ ...value, minutes: parseInt(e.target.value, 10) })}
+        style={style}
+        value={value.minutes}
       >
         {genArray(0, 59).map(min =>
           <option key={min} value={min}>{leftPadTimeUnit(min)}</option>)}
       </select>
-      { select24Hours ? null : renderAmPm({ value, onChange, submitting }) }
+      { select24Hours ? null : renderAmPm({ noStyle, onChange, submitting, value }) }
     </div>
   );
 };
 
 InputTime.propTypes = {
-  select24Hours: PropTypes.bool,
   input: PropTypes.shape({
+    onChange: PropTypes.func.isRequired,
     value: PropTypes.oneOfType([
       PropTypes.shape({
         hours: PropTypes.number.isRequired,
@@ -91,11 +138,12 @@ InputTime.propTypes = {
       }),
       PropTypes.string,
     ]),
-    onChange: PropTypes.func.isRequired,
   }).isRequired,
   meta: PropTypes.shape({
     submitting: PropTypes.bool,
   }),
+  noStyle: PropTypes.bool,
+  select24Hours: PropTypes.bool,
 };
 
 InputTime.defaultProps = {