+++ InputTime/story.jsx
@@ -12,16 +12,35 @@ storiesOf('InputTime')
   .add('default', () => (
     <InputTime
       input={{
-        value: '',
         onChange: action('on-change'),
+        value: '',
       }}
+    />
+  ))
+  .add('noStyle', () => (
+    <InputTime
+      input={{
+        onChange: action('on-change'),
+        value: '',
+      }}
+      noStyle
+    />
+  ))
+  .add('noStyle with 24 hour selection', () => (
+    <InputTime
+      input={{
+        onChange: action('on-change'),
+        value: '',
+      }}
+      noStyle
+      select24Hours
     />
   ))
   .add('with 24 hour selection', () => (
     <InputTime
       input={{
-        value: '',
         onChange: action('on-change'),
+        value: '',
       }}
       select24Hours
     />
@@ -29,33 +48,33 @@ storiesOf('InputTime')
   .add('with value set', () => (
     <InputTime
       input={{
+        onChange: action('on-change'),
         value: {
           hours: 10,
           minutes: 32,
         },
-        onChange: action('on-change'),
       }}
     />
   ))
   .add('with afternoon value set', () => (
     <InputTime
       input={{
+        onChange: action('on-change'),
         value: {
           hours: 11,
           minutes: 32,
         },
-        onChange: action('on-change'),
       }}
     />
   ))
   .add('with afternoon value set and 24 hour selection', () => (
     <InputTime
       input={{
+        onChange: action('on-change'),
         value: {
           hours: 16,
           minutes: 33,
         },
-        onChange: action('on-change'),
       }}
       select24Hours
     />
@@ -63,11 +82,11 @@ storiesOf('InputTime')
   .add('with form submitting', () => (
     <InputTime
       input={{
+        onChange: action('on-change'),
         value: {
           hours: 16,
           minutes: 33,
         },
-        onChange: action('on-change'),
       }}
       meta={{
         submitting: true,