+++ Link/test.jsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import LinkStateless from '../LinkStateless';
+import Link from './index';
+
+describe('Link', () => {
+  it('should toggle hovered prop when mouseEnter and mouseLeave are triggered', () => {
+    const wrapper = mount(
+      <Link href={'localhost:8000'}>A Link</Link>,
+    );
+    const link = wrapper
+      .find(LinkStateless);
+    link.simulate('mouseEnter');
+    expect(link.props().hovered)
+      .toBe(true);
+    link.simulate('mouseLeave');
+    expect(link.props().hovered)
+      .toBe(false);
+  });
+
+  it('should toggle focused prop when focus and blur are triggered', () => {
+    const wrapper = mount(
+      <Link href={'localhost:8000'}>A Link</Link>,
+    );
+    const link = wrapper
+      .find(LinkStateless);
+    link.simulate('focus');
+    expect(link.props().focused)
+      .toBe(true);
+    link.simulate('blur');
+    expect(link.props().focused)
+      .toBe(false);
+  });
+
+  it('should toggle focused prop when focus and blur are triggered', () => {
+    /* eslint-disable react/prop-types */
+    const HoverableFocusableThing = ({ hovered, focused }) =>
+      <div>{hovered} - {focused}</div>;
+    /* eslint-enable react/prop-types */
+    const wrapper = mount(
+      <Link>
+        <HoverableFocusableThing />
+      </Link>,
+    );
+    const link = wrapper
+      .find(LinkStateless);
+    link.simulate('focus');
+    expect(link.props().focused)
+      .toBe(true);
+    link.simulate('blur');
+    expect(link.props().focused)
+      .toBe(false);
+  });
+});