+++ index.js
@@ -18,4 +18,6 @@ export NavBar from './NavBar';
 export MultipleImages from './MultipleImages';
 export Overlay from './Overlay';
 export Popover from './Popover';
+export SectionHeader from './SectionHeader';
+export Select from './Select';
 export Text from './Text';