+++ style/font.js
@@ -1,0 +1,13 @@
+export const fontFamily = '"Open Sans" sans-serif';
+
+export const fontSize = '1rem';
+export const fontSizeLarge = '2rem';
+export const fontSizeMini = '0.9rem';
+export const fontSizeSmall = '0.8rem';
+export const fontSizeExtraSmall = '0.7rem';
+
+export const fontWeightBold = 600;
+export const fontWeight = 400;
+export const fontWeightThin = 300;
+
+export const lineHeight = 1.5;