+++ CHANGELOG.md
@@ -1,48 +1,114 @@
 # Changelog
 
+## 0.5.10 (May 26, 2017)
+
+- Add `noStyle` config to `InputTimes`
+
+## 0.5.9 (May 25, 2017)
+
+- Fix a slight browser UI bug that occurs when a `select` element is the child of a `div`
+
+## 0.5.8 (May 25, 2017)
+
+- Remove `icon` config from `Button`
+
+## 0.5.7 (May 24, 2017)
+
+- Replace `background` with `backgroundColor` on all `icon` config states
+
+## 0.5.6 (May 24, 2017)
+
+- Replace `backgroundColor` with `background` on `icon` config hover state
+
+## 0.5.5 (May 24, 2017)
+
+- Add `iconHovered` state to `ButtonStateless` `icon` config
+
+## 0.5.4 (May 24, 2017)
+
+- Remove `lineHeight` from `ButtonStateless` and replace with `icon` config
+
+## 0.5.3 (May 24, 2017)
+
+- Apply `lineHeight: 0` to `ButtonStateless`
+
+## 0.5.2 (May 23, 2017)
+
+- `Button` manages own focus and hover state
+- `Link` manages own focus and hover state
+
+## 0.5.1 (May 19, 2017)
+
+- Fix `Button` noStyle
+- Make `Link` focusable/blurable
+
+## 0.5.0 (May 19, 2017)
+
+- Migrate to Inline Styles
+
+## 0.4.24 (May 18, 2017)
+
+- Replace `height` CSS property on `SectionHeader` with `padding-top` and `padding-bottom`
+
+## 0.4.23 (May 17, 2017)
+
+- Add `Select`
+
+## 0.4.22 (May 17, 2017)
+
+- Update `SectionHeader` CSS `flex-direction` add `align-items` with `center`
+
+## 0.4.21 (May 17, 2017)
+
+- Add `SectionHeader`
+
+## 0.4.20 (May 16, 2017)
+
+- Add `outerSpace` config to `Text`
+
 ## 0.4.19 (April 26, 2017)
 
 - All Input-* handle submitting state
 
 ## 0.4.18 (April 26, 2017)
 
-- Add `color` and `reducedPadding` configurations to `Card`
+- Add `color` and `reducedPadding` configs to `Card`
 
 ## 0.4.17 (April 25, 2017)
 
-- Add `bottom` and `right` configurations to `Popover`
+- Add `bottom` and `right` configs to `Popover`
 
 ## 0.4.16 (April 19, 2017)
 
-- Bugfix on `InputDate` and `InputTime` components
+- Bugfix on `InputDate` and `InputTime`
 
 ## 0.4.15 (April 17, 2017)
 
-- Add `InputTime` component
+- Add `InputTime`
 
 ## 0.4.14 (April 17, 2017)
 
-- Add `newTab` and `unstyled` config to `LinkifiedText` component
+- Add `newTab` and `unstyled` configs to `LinkifiedText`
 
 ## 0.4.13 (April 17, 2017)
 
-- Add `newTab` config to `Link` component
+- Add `newTab` config to `Link`
 
 ## 0.4.12 (April 12, 2017)
 
-- Add `cursor: default;` to `disabled` config of `Button` component
+- Add `cursor: default;` to `disabled` config of `Button`
 
 ## 0.4.11 (April 11, 2017)
 
-- Add `Input`, `InputEmail` and `InputPassword` components
+- Add `Input`, `InputEmail` and `InputPassword`
 
 ## 0.4.10 (April 11, 2017)
 
-- Add `IdTag` component
+- Add `IdTag`
 
 ## 0.4.9 (April 6, 2017)
 
-- Add `Popover` component
+- Add `Popover`
 
 ## 0.4.8 (April 6, 2017)
 
@@ -63,15 +129,15 @@
 
 ## 0.4.4 (March 24, 2017)
 
-- Set `Button`'s `margin` property to `0`
+- Set `Button`'s `margin` to `0`
 
 ## 0.4.3 (March 21, 2017)
 
-- Update `MultipleImages` to have a `rounded` flag to set a border-radius on itself
+- Update `MultipleImages` to have a `rounded` flag to set a `border-radius` on itself
 
 ## 0.4.2 (March 20, 2017)
 
-- Add export of `MultipleImages` component to `index.js`
+- Add export of `MultipleImages` to `index.js`
 
 ## 0.4.1 (March 16, 2017)
 
@@ -81,7 +147,7 @@
 
 ## 0.4.0 (March 16, 2017)
 
-- Add `MultipleImages` component
+- Add `MultipleImages`
 
 ## 0.3.0 (March 14, 2017)
 
@@ -89,7 +155,7 @@
 
 ## 0.2.2 (March 8, 2017)
 
-- Add blue color config to `Text` component
+- Add blue color config to `Text`
 
 ## 0.2.1 (February 23, 2017)
 
@@ -98,7 +164,7 @@
 
 ## 0.2.0 (February 23, 2017)
 
-- Remove our icon font dependancy in replace of our new `Icon` component
+- Remove our icon font dependancy in replace of our new `Icon`
 
 ## 0.1.25 (February 22, 2017)
 
@@ -111,7 +177,7 @@
 ## 0.1.23 (February 22, 2017)
 
 - Add and use SVG `Icons`
-- Add `Overlay` component
+- Add `Overlay`
 
 ## 0.1.19 (February 16, 2017)
 
@@ -122,7 +188,7 @@ Bring these back:
 
 ## 0.1.18 (February 16, 2017)
 
-- Add `minHeight` and `minWidth` properties to `Image`
+- Add `minHeight` and `minWidth` to `Image`
 
 ## 0.1.16 (February 13, 2017)
 
@@ -140,7 +206,7 @@ Bring these back:
 
 ## 0.1.11 (February 3, 2017)
 
-- Added `hovered` prop to `Card` component
+- Added `hovered` prop to `Card`
 - Remove `layout.css` file and references
 
 ## 0.1.3 (January 10, 2017)
@@ -149,7 +215,7 @@ Bring these back:
 
 ## 0.1.2 (January 10, 2017)
 
-- Remove `margin-bottom` property from the `Card` component
+- Remove `margin-bottom` from `Card`
 
 ## 0.1.1 (January 9, 2017)